Dropbox.munki: Fix postinstall and install-check
The postinstall script needs to set the setuid bit after chowning. Recent macOS versions appear to remove a setuid bit when the owner changes.

The presence of any Dropbox version was considered as sufficient criterion for Dropbox being up-to-date. This meant that we relied on Dropbox's auto-updater to download newer versions to the user's home folder. Unfortunately, this only worked as long as DropboxHelperInstaller remained compatible between Dropbox versions.